requirements.txt
Flask==3.0.3
APScheduler
requests==2.32.3
beautifulsoup4
python-telegram-bot
